Git wants e to commit, I want to sleep. Take me sweet void.
