+++ src/components/appbar/ApplicationBar.tsx
@@ -1,14 +1,9 @@
 import AppBar from '@mui/material/AppBar'
-import Box from '@mui/material/Box'
 import Toolbar from '@mui/material/Toolbar'
-import Typography from '@mui/material/Typography'
 import Button from '@mui/material/Button'
-import IconButton from '@mui/material/IconButton'
-import MenuIcon from '@mui/icons-material/Menu'
 import { LogoutOptions, useAuth0 } from '@auth0/auth0-react'
-import TemporaryDrawer from '../drawer/ApplicationDrawer'
-import React, { useState } from 'react'
-import { useLocation } from 'react-router-dom'
+import { useNavigate } from 'react-router-dom'
+import { Grid } from '@mui/material'
 
 const logoutOption: LogoutOptions = {
   logoutParams: {
@@ -16,52 +11,36 @@ const logoutOption: LogoutOptions = {
   }
 }
 
-function stringSwitch(inputStr: string): string {
-  if (/^\/group\/\d+$/.test(inputStr)) {
-    return 'Group with ID'
-  }
-
-  switch (inputStr) {
-    case '/home':
-      return 'Home'
-    case '/groups':
-      return 'Groups'
-    case '/bankaccounts':
-      return 'Bank Accounts'
-    default:
-      return 'Unknown'
-  }
-}
-
 export default function ApplicationBar() {
   const { logout } = useAuth0()
-  const [drawerIsOpen, setDrawerIsOpen] = useState<boolean>(false)
-
-  const location = useLocation()
-  const label = stringSwitch(location.pathname)
+  const navigate = useNavigate()
 
   const logoutProcess = () => {
     logout(logoutOption)
   }
 
   return (
-    <React.Fragment>
-      <TemporaryDrawer open={drawerIsOpen} onClose={() => setDrawerIsOpen(false)} />
-      <Box sx={{ flexGrow: 1 }}>
-        <AppBar position="static">
-          <Toolbar>
-            <IconButton size="large" edge="start" color="inherit" aria-label="menu" sx={{ mr: 2 }} onClick={() => setDrawerIsOpen(true)}>
-              <MenuIcon />
-            </IconButton>
-            <Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>
-              {label}
-            </Typography>
+    <AppBar position="static">
+      <Toolbar>
+        <Grid container justifyContent={'space-between'}>
+          <Grid item>
+            <Button color="inherit" onClick={() => navigate('/home')}>
+              Home
+            </Button>
+            <Button color="inherit" onClick={() => navigate('/groups')}>
+              Groups
+            </Button>
+            <Button color="inherit" onClick={() => navigate('/bankaccounts')}>
+              Bank Account
+            </Button>
+          </Grid>
+          <Grid item>
             <Button color="inherit" onClick={logoutProcess}>
               Logout
             </Button>
-          </Toolbar>
-        </AppBar>
-      </Box>
-    </React.Fragment>
+          </Grid>
+        </Grid>
+      </Toolbar>
+    </AppBar>
   )
 }
