+++ src/pages/groups/GroupsPage.tsx
@@ -1,14 +1,10 @@
-import { useGroupDeleteGroup, useGroupGetGroups } from '@/tools/hooks/apihooks/groupapihook'
+import { useGroupGetGroups } from '@/tools/hooks/apihooks/groupapihook'
 import { useUserGetCurrentUserInfo } from '@/tools/hooks/apihooks/userapihook'
 import { Fab, Paper, Table, TableBody, TableCell, TableContainer, TableHead, TableRow } from '@mui/material'
 import React, { useEffect, useState } from 'react'
 import AddIcon from '@mui/icons-material/Add'
-import { GroupDTO } from '@jbwittner/bankwiz_openapi-client-fetch'
-import ValidationDialog from '@/components/dialog/ValidationDialog'
 import GroupCreationDialog from './components/GroupCreationDialog'
 import { GroupeLine } from './components/GroupLine'
-import useConfirmationModal from '@/tools/hooks/component/modalhook'
-import GroupUsersDialog from './components/GroupUsersDialog'
 
 const fabSx = {
   margin: 0,
@@ -20,25 +16,9 @@ const fabSx = {
 }
 
 export const GroupsPage: React.FC = () => {
-  const {
-    isOpen: isGroupDeletionModalOpen,
-    data: groupToDelete,
-    openModal: openGroupDeletionModal,
-    closeModal: closeGroupDeletionModal,
-    confirmAction: confirmGroupDeletion
-  } = useConfirmationModal<GroupDTO>()
-
-  const {
-    isOpen: isGroupUserManagementModalOpen,
-    data: groupWithUsers,
-    openModal: openGroupUserManagementModal,
-    closeModal: closeGroupUserManagementModal
-  } = useConfirmationModal<GroupDTO>()
-
   const [modalCreateIsOpen, setModalCreateIsOpen] = useState(false)
   const { groupsDTO, getGroups } = useGroupGetGroups()
   const { userDTO, getCurrentUserInfo } = useUserGetCurrentUserInfo()
-  const { deleteGroup } = useGroupDeleteGroup()
 
   useEffect(() => {
     getGroups()
@@ -50,24 +30,9 @@ export const GroupsPage: React.FC = () => {
     setModalCreateIsOpen(false)
   }
 
-  const confirmDeleteCallback = async (data: GroupDTO) => {
-    await deleteGroup(data.groupId)
-    await getGroups()
-  }
-
   return (
     <React.Fragment>
       <GroupCreationDialog open={modalCreateIsOpen} onValid={onCloseModalCreation} onCancel={() => setModalCreateIsOpen(false)} />
-      <ValidationDialog
-        titleDialog={'Deletion confirmation'}
-        textDialog={'You will delete the group ' + groupToDelete?.groupName}
-        open={isGroupDeletionModalOpen}
-        onCancel={closeGroupDeletionModal}
-        onValid={() => confirmGroupDeletion(confirmDeleteCallback)}
-      />
-      {groupWithUsers && userDTO && (
-        <GroupUsersDialog open={isGroupUserManagementModalOpen} group={groupWithUsers} currentUser={userDTO} onClose={closeGroupUserManagementModal} />
-      )}
       <TableContainer
         component={Paper}
         sx={{
@@ -87,19 +52,7 @@ export const GroupsPage: React.FC = () => {
               <TableCell align="center">Actions</TableCell>
             </TableRow>
           </TableHead>
-          <TableBody>
-            {groupsDTO &&
-              userDTO &&
-              groupsDTO.map(group => (
-                <GroupeLine
-                  key={group.groupId}
-                  groupDTO={group}
-                  userDTO={userDTO}
-                  onClickDelete={openGroupDeletionModal}
-                  onClickUsers={openGroupUserManagementModal}
-                />
-              ))}
-          </TableBody>
+          <TableBody>{groupsDTO && userDTO && groupsDTO.map(group => <GroupeLine key={group.groupId} groupDTO={group} userDTO={userDTO} />)}</TableBody>
         </Table>
       </TableContainer>
       <Fab color="primary" aria-label="add" sx={fabSx} onClick={() => setModalCreateIsOpen(true)}>
